publish Endpoint bug
Expect : Publish statd data to application insight.
Actual: Failed to publish. Error: The document is moved. The endpoint 'http' protocol it was publishing throwing a redirection response.
Change : upgrade the applicationinsights dependency to 0.17.2 which change the protocol to https.
The application endpoint doesn't accept http request and the fix was added with release 0.16.0.